moving util funcs to util.py
